+++ srv/generate_plot.py
@@ -2,20 +2,34 @@ import io
 import datetime
 
 import numpy as np
-import matplotlib as mpl
-import matplotlib.pyplot as plt
+import matplotlib.dates as mdates
 
 from requests import get
 from urllib.request import urlopen
 from scipy.interpolate import griddata
 
-from bokeh.plotting import figure
+from bokeh.plotting import figure, output_file, save, show
 from bokeh.embed import components
 from bokeh.models import RangeTool, Range1d, CustomJSTickFormatter
 from bokeh.layouts import column
-from bokeh.palettes import Sunset8
+from bokeh import palettes
+from bokeh import __version__ as bokeh_version
+from bokeh.models import DatetimeTicker  # , DatetimeTickFormatter
 
-template_html = """
+# ----------------------------------------------------------------------------------------
+
+# save only plot to single html file?
+SAVE_STATIC = False
+
+# only show the plot, don't save it?
+SHOW = False
+
+# if the specified palette is not available, it will fall back to 'Spectral8'
+COLOR_PALETTE = "RdBu"
+
+# ----------------------------------------------------------------------------------------
+
+template_html = f"""
 <!DOCTYPE html>
 <html lang="en">
   <head>
@@ -25,18 +39,40 @@ template_html = """
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-3.4.0.min.js"></script>
-    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-widgets-3.4.0.min.js"></script>
+    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-{bokeh_version}.min.js"></script>
+    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-widgets-{bokeh_version}.min.js"></script>
     <script type="text/javascript">
         Bokeh.set_log_level("info");
     </script>
   </head>
   <body>
-    {{TEMPLATE_DIV}}
+    {{{{TEMPLATE_DIV}}}}
   </body>
   <script src="plot.js"></script>
 </html>
 """
+
+# ----------------------------------------------------------------------------------------
+
+
+def get_palette(palette_name, n_colors=8):
+    ### Claude 3.7 ###
+    # Try to get the palette with the specified number of colors
+    try:
+        palette_full_name = f"{palette_name}{n_colors}"
+        return getattr(palettes, palette_full_name)
+    except AttributeError:
+        # If that exact number isn't available, try to get a larger palette and slice it
+        for size in [9, 10, 11, 12, 256]:  # Common palette sizes
+            try:
+                palette_full_name = f"{palette_name}{size}"
+                return getattr(palettes, palette_full_name)[:n_colors]
+            except AttributeError:
+                continue
+
+        # If we can't find the palette, fall back to a default
+        print(f"Palette {palette_name} not found. Using Spectral8 instead.")
+        return palettes.Spectral8
 
 
 def download(url, file_name):
@@ -89,7 +125,7 @@ def read_singapore(nmonth, nyear):
             count = count + 1
     pressure = [100, 90, 80, 70, 60, 50, 45, 40, 35, 30, 25, 20, 15, 12, 10]
     altitude = -7 * np.log(np.array(pressure) / 1013.25)
-    fds = list(mpl.dates.date2num(date))
+    fds = list(mdates.date2num(date))
     fds = np.array(fds)
     return np.array(dats).T[::-1], fds, pressure, altitude
 
@@ -166,7 +202,7 @@ def read_qbo():
                     )
                 )
 
-    fds = mpl.dates.date2num(date)
+    fds = mdates.date2num(date)
     #
     up = np.array(
         list(
@@ -193,15 +229,19 @@ def main():
     tnmoth = np.shape(up1)[1]  # total number of months
     up = np.zeros([15, tnmoth]) * np.nan
     nmonth = np.shape(up1)[1] % 12  # number of months above a whole year
-    nyear = mpl.dates.num2date(fds1[-1]).year
+    nyear = mdates.num2date(fds1[-1]).year
     if nmonth == 0:
         nyear += 1
     print(nmonth, nyear)
     up2, fds2, pressure, altitude = read_singapore(nmonth, nyear)
     fds = fds1
+    print(fds[-fds2.size], fds[-fds2.size + 1])
+    print(fds2[0])
+    print(np.size(fds1), np.size(fds2))
     fds[-fds2.size :] = fds2
     fds = np.array(fds)
-    print(fds[-1])
+    print(np.size(fds), fds[-1], fds1[-1], fds2[-1])
+    print(mdates.num2date(fds2[-1]))
     up1 = up1 * 1.0
     up1[up1 < -10000] = np.nan
     up[3, :] = up1[0, :]
@@ -225,7 +265,7 @@ def main():
     # print(points1.shape)
     X, Y = np.meshgrid(fds, pressure)
     u = griddata(points1, uu, (X, Y), method="linear")
-    nrows = 7
+    # nrows = 7
 
     contour_levels = range(-40, 45, 5)
     deltat = int((np.nanmax(fds) - fds[0]) / 1) + 1
@@ -237,22 +277,27 @@ def main():
 
     axz = u[:, np.nanargmin(abs(fds - by0)) : 1 + np.nanargmin(abs(fds - by1))]
 
-    # axt = [t.strftime("%m/%Y") for t in mpl.dates.num2date(axt)]
+    # axt = [t.strftime("%m/%Y") for t in mdates.num2date(axt)]
 
     p = figure(
         width=1200,
         height=500,
         y_axis_type="log",
-        y_range=(max(pressure), min(pressure)),
-        x_range=(axt[0], axt[100]),
+        y_range=(max(pressure), min(pressure)),  # type: ignore
+        x_range=(axt[0], axt[100]),  # type: ignore
         y_axis_label="Pressure [hPa]",
         # title="Quasi-Biennial-Oscillation (QBO)"
     )
-    p.title.text_font_size = "25px"
-    p.title.align = "center"
+    p.title.text_font_size = "25px"  # type: ignore
+    p.title.align = "center"  # type: ignore
 
     contour_renderer = p.contour(
-        axt, pressure, axz, contour_levels, fill_color=Sunset8, line_color="black"
+        axt + 7 * 30,
+        pressure,
+        axz,
+        contour_levels,
+        fill_color=get_palette(COLOR_PALETTE),  # Sunset8,
+        line_color="black",
     )
 
     select = figure(
@@ -270,57 +315,129 @@ def main():
     range_tool.overlay.fill_alpha = 0.4
 
     select.contour(
-        axt,
+        axt + 5 * 30 * 12 + 4 * 30 + 15,
         pressure,
         axz,
         contour_levels,
         line_alpha=0.1,
-        fill_color=Sunset8,
+        fill_color=get_palette(COLOR_PALETTE),  # Sunset8,
         line_color="black",
     )
     select.add_tools(range_tool)
 
-    p.xaxis[0].formatter = CustomJSTickFormatter(
+    # # Set yearly intervals for the top plot
+    # p.xaxis.ticker = DatetimeTicker(desired_num_ticks=22)  # Adjust for more control over spacing
+
+    # p.xaxis[0].formatter = CustomJSTickFormatter(
+    #     code="""
+    # const milliseconds = tick * 24 * 60 * 60 * 1000;
+
+    # // Create a Date object using the calculated milliseconds
+    # const date = new Date(milliseconds);
+
+    # // Extract the month and year from the Date object
+    # const month = (date.getUTCMonth() + 1).toString().padStart(2, '0'); // Month is 0-based
+    # const year = date.getUTCFullYear();
+
+    # // Format the result as MM/YYYY
+    # return `${month}/${year}`;
+    # """
+    # )
+    # # Set 10-year intervals for the bottom plot
+    # select.xaxis.ticker = DatetimeTicker(desired_num_ticks=15)  # Larger gap for decades
+
+    # select.xaxis[0].formatter = CustomJSTickFormatter(
+    #     code="""
+    # const milliseconds = tick * 24 * 60 * 60 * 1000;
+
+    # // Create a Date object using the calculated milliseconds
+    # const date = new Date(milliseconds);
+
+    # // Extract the month and year from the Date object
+    # const month = (date.getUTCMonth() + 1).toString().padStart(2, '0'); // Month is 0-based
+    # const year = date.getUTCFullYear();
+
+    # // Format the result as MM/YYYY
+    # return `${month}/${year}`;
+    # """
+    # )
+
+    # # Top plot: yearly labels starting from January 1st
+    # p.xaxis.formatter = CustomJSTickFormatter(
+    # code="""
+    # const date = new Date(tick * 24 * 60 * 60 * 1000);
+    # date.setUTCMonth(0);
+    # date.setUTCDate(1);
+    # return `${date.getUTCFullYear()}`;
+    # """
+    # )
+
+    # # Bottom plot: 10-year intervals with clean formatting
+    # select.xaxis.formatter = CustomJSTickFormatter(
+    # code="""
+    # const date = new Date(tick * 24 * 60 * 60 * 1000);
+    # const year = date.getUTCFullYear();
+    # if (year % 10 === 0) {  // Show only on 10-year intervals
+    #     return `${year}`;
+    # } else {
+    #     return "";
+    # }
+    # """
+    # )
+
+    # Top Plot (Align the labels correctly)
+
+    p.xaxis.formatter = CustomJSTickFormatter(  # type: ignore
         code="""
-    const milliseconds = tick * 24 * 60 * 60 * 1000;
-
-    // Create a Date object using the calculated milliseconds
-    const date = new Date(milliseconds);
-
-    // Extract the month and year from the Date object
-    const month = (date.getUTCMonth() + 1).toString().padStart(2, '0'); // Month is 0-based
-    const year = date.getUTCFullYear();
-
-    // Format the result as MM/YYYY
-    return `${month}/${year}`;
-    """
+        const date = new Date(tick * 24 * 60 * 60 * 1000);  // Convert from date2num to milliseconds
+        const year = date.getUTCFullYear();  // Extract the year
+        return `${year}`;  // Return the year
+        """
     )
-    select.xaxis[0].formatter = CustomJSTickFormatter(
+
+    # Bottom Plot (10-Year Interval Labels)
+    select.xaxis.formatter = CustomJSTickFormatter(  # type: ignore
         code="""
-    const milliseconds = tick * 24 * 60 * 60 * 1000;
-
-    // Create a Date object using the calculated milliseconds
-    const date = new Date(milliseconds);
-
-    // Extract the month and year from the Date object
-    const month = (date.getUTCMonth() + 1).toString().padStart(2, '0'); // Month is 0-based
-    const year = date.getUTCFullYear();
-
-    // Format the result as MM/YYYY
-    return `${month}/${year}`;
-    """
+        const date = new Date(tick * 24 * 60 * 60 * 1000);  // Convert from date2num to milliseconds
+        const year = date.getUTCFullYear();  // Extract the year
+        const roundedYear = Math.floor(year / 10) * 10;  // Round to nearest decade
+        return `${roundedYear}`;  // Return the decade
+        """
     )
 
-    select.yaxis.major_label_text_font_size = "0pt"
+    p.xaxis.ticker.num_minor_ticks = 12  # type: ignore
+    select.xaxis.ticker.num_minor_ticks = 10  # type: ignore
+    # select.xaxis.formatter = CustomJSTickFormatter(
+    #    code="""
+    #    const date = new Date(tick * 24 * 60 * 60 * 1000);  // Convert from date2num to milliseconds
+    #    const year = date.getUTCFullYear();  // Extract the year
+    #    // Round to the nearest decade
+    #    const roundedYear = Math.floor(year / 10) * 10;
+    #    // Shift the first label by half a decade (5 years)
+    #    const shiftedYear = roundedYear + 5;
+    #    return `${shiftedYear}`;  // Return the shifted year
+    #    """
+    # )
 
-    select.yaxis.major_tick_line_color = None  # turn off y-axis major ticks
-    select.yaxis.minor_tick_line_color = None  # turn off y-axis minor ticks
+    # Adjust ticker for bottom plot to show yearly ticks (desired_num_ticks adjusted)
+    select.xaxis.ticker = DatetimeTicker(  # type: ignore
+        desired_num_ticks=12
+    )  # Adjust to suit your range
 
-    p.axis.axis_label_text_font_size = "20px"
-    select.axis.axis_label_text_font_size = "20px"
-    p.axis.major_label_text_font_size = "15px"
-    select.axis.major_label_text_font_size = "15px"
-    select.yaxis.major_label_text_font_size = "0px"
+    select.yaxis.major_label_text_font_size = "0pt"  # type: ignore
+
+    select.yaxis.major_tick_line_color = (  # type: ignore
+        None  # turn off y-axis major ticks
+    )
+    select.yaxis.minor_tick_line_color = (  # type: ignore
+        None  # turn off y-axis minor ticks
+    )
+
+    p.axis.axis_label_text_font_size = "20px"  # type: ignore
+    select.axis.axis_label_text_font_size = "20px"  # type: ignore
+    p.axis.major_label_text_font_size = "15px"  # type: ignore
+    select.axis.major_label_text_font_size = "15px"  # type: ignore
+    select.yaxis.major_label_text_font_size = "0px"  # type: ignore
 
     altitude = -7 * np.log(np.array(pressure) / 1013.25)
 
@@ -333,10 +450,18 @@ def main():
     p.add_layout(colorbar, "below")
 
     select.title = "Move/resize slider by dragging"
-    select.title.text_font_size = "15px"
+    select.title.text_font_size = "15px"  # type: ignore
 
-    # show(column(p, select))
-    # show(p)
+    if SHOW:
+        show(column(p, select))
+        # show(p)
+        return None, None
+
+    if SAVE_STATIC:
+        output_file(filename="plot_only.html", title="Static HTML file")
+        save(p)
+        return None, None
+
     return components(column(p, select))
 
 
@@ -344,15 +469,16 @@ if __name__ == "__main__":
 
     script, div = main()
 
-    print(div)
-    updated_html = template_html.replace("{{TEMPLATE_DIV}}", div)
+    if all(x is not None for x in (script, div)):
+        # print(div)
+        updated_html = template_html.replace("{{TEMPLATE_DIV}}", div)  # type: ignore
 
-    with open("plot.js", "w") as file:
-        file.write(
-            script.replace('<script type="text/javascript">', "").replace(
-                "</script>", ""
+        with open("plot.js", "w") as file:
+            file.write(
+                script.replace('<script type="text/javascript">', "").replace(  # type: ignore
+                    "</script>", ""
+                )
             )
-        )
 
-    with open("plot.html", "w") as file:
-        file.write(updated_html)
+        with open("plot.html", "w") as file:
+            file.write(updated_html)
